feat: today posts count
